revise main.js, add index.js file to run the server from, and use dirPublic as argument for express static method
