+++ express-static/public/index.html
@@ -4,10 +4,10 @@
   <meta charset="UTF-8">
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link rel="stylesheet" href="/styles.css">
+  <link rel="stylesheet" href="./styles.css">
   <title>Express Static</title>
 </head>
 <body>
-  <script src="/main.js"></script>
+  <script src="./main.js"></script>
 </body>
 </html>
