+++ express-static/public/main.js
@@ -1,17 +1,2 @@
-const express = require('express');
-const path = require('path');
-
-const app = express();
-
-const dirPublic = path.join(__dirname, 'public');
 // eslint-disable-next-line no-console
-console.log(dirPublic);
-
-const staticPublic = express.static('public');
-
-app.use(staticPublic);
-
-app.listen(3000, () => {
-  // eslint-disable-next-line no-console
-  console.log('listening on port 3000');
-});
+console.log('Static Middleware mounted successfully');
